+++ src/main/java/com/tsp/new_tsp_front/api/model/negotiation/service/impl/FrontNegotiationJpaRepository.java
@@ -179,7 +179,7 @@ public class FrontNegotiationJpaRepository {
      * 5. 작성일       : 2022. 09. 11.
      * </pre>
      */
-    public Integer deleteModelNegotiation(Integer idx) {
+    public Long deleteModelNegotiation(Long idx) {
         em.remove(em.find(FrontNegotiationEntity.class, idx));
         em.flush();
         em.clear();
